Revert splitting post-view-data and post-data-editor controllers into different file
Makes for easier diff'ing. Sorry @rowasc
